==> 5_card_feat_eng.ipynb <==
import numpy as np
import pandas as pd
from warnings import filterwarnings as f_w
f_w('ignore')
pd.options.display.max_columns = 999
pd.options.display.max_rows = 500
import gc

table = 'card' 
locator = pd.read_excel('locator.xlsx', index_col = 0)

card = pd.read_csv(locator.loc[table, 'initial'])

card = card.sort_values(by = ['SK_ID_PREV', 'MONTHS_BALANCE'])

card['ATM_AMT_RATIO'] = card.eval('AMT_DRAWINGS_ATM_CURRENT / AMT_DRAWINGS_CURRENT')
card['POS_AMT_RATIO'] = card.eval('AMT_DRAWINGS_POS_CURRENT / AMT_DRAWINGS_CURRENT')
card['OTH_AMT_RATIO'] = card.eval('AMT_DRAWINGS_OTHER_CURRENT / AMT_DRAWINGS_CURRENT')
card['ATM_CNT_RATIO'] = card.eval('CNT_DRAWINGS_ATM_CURRENT / CNT_DRAWINGS_CURRENT')
card['POS_CNT_RATIO'] = card.eval('CNT_DRAWINGS_POS_CURRENT / CNT_DRAWINGS_CURRENT')
card['OTH_CNT_RATIO'] = card.eval('CNT_DRAWINGS_OTHER_CURRENT / CNT_DRAWINGS_CURRENT')
card['MEAN_DRAWING'] = card.eval('AMT_DRAWINGS_CURRENT / CNT_DRAWINGS_CURRENT')
card['CUR_PAYM_RATIO'] = card.eval('AMT_PAYMENT_CURRENT / AMT_PAYMENT_TOTAL_CURRENT')
card['RECIEVPRIN_RATIO'] = card.eval('AMT_RECEIVABLE_PRINCIPAL / AMT_TOTAL_RECEIVABLE')
card['RECIEV_RATIO'] = card.eval('AMT_RECIVABLE / AMT_TOTAL_RECEIVABLE')
card['BALANCE_LIMIT'] = card.eval('AMT_BALANCE / AMT_CREDIT_LIMIT_ACTUAL')
card['DRAWING_LIMIT'] = card.eval('AMT_DRAWINGS_CURRENT / AMT_CREDIT_LIMIT_ACTUAL')
card['PAYMENT_LIMIT'] = card.eval('AMT_PAYMENT_TOTAL_CURRENT / AMT_CREDIT_LIMIT_ACTUAL')
card['DRAWING_BALANCE'] = card.eval('AMT_DRAWINGS_CURRENT / AMT_BALANCE')
card['PAYMENT_BALANCE'] = card.eval('AMT_PAYMENT_TOTAL_CURRENT / AMT_BALANCE')
card['RECIVABLE_LIMIT'] = card.eval('AMT_TOTAL_RECEIVABLE / AMT_CREDIT_LIMIT_ACTUAL')
card['RECIVABLE_BALANCE'] = card.eval('AMT_TOTAL_RECEIVABLE / AMT_BALANCE')
card['PAYMENT_RECIVABLE'] = card.eval('AMT_PAYMENT_TOTAL_CURRENT / AMT_TOTAL_RECEIVABLE')
card['MININS_LIMIT'] = card.eval('AMT_INST_MIN_REGULARITY / AMT_CREDIT_LIMIT_ACTUAL')
card['MININS_BALANCE'] = card.eval('AMT_INST_MIN_REGULARITY / AMT_BALANCE')
card['MININS_RECIVABLE'] = card.eval('AMT_INST_MIN_REGULARITY / AMT_TOTAL_RECEIVABLE')

card = card.replace({np.inf:np.nan,
                    -np.inf:np.nan})

aggregations = {
    'MONTHS_BALANCE': ['min', 'size'],
    'AMT_BALANCE': ['mean', 'std', 'last'],
    'AMT_CREDIT_LIMIT_ACTUAL': ['mean', 'std', 'last'],
    'AMT_DRAWINGS_ATM_CURRENT': ['mean', 'std'],
    'AMT_DRAWINGS_CURRENT': ['mean', 'std'],
    'AMT_DRAWINGS_OTHER_CURRENT': ['mean', 'std'],
    'AMT_DRAWINGS_POS_CURRENT': ['mean', 'std'],
    'AMT_INST_MIN_REGULARITY': ['mean', 'std'],
    'AMT_PAYMENT_CURRENT': ['mean', 'std'],
    'AMT_PAYMENT_TOTAL_CURRENT': ['mean', 'std'],
    'AMT_RECEIVABLE_PRINCIPAL': ['mean', 'std'],
    'AMT_RECIVABLE': ['mean', 'std'],
    'AMT_TOTAL_RECEIVABLE': ['mean', 'std', 'last'],
    'CNT_DRAWINGS_ATM_CURRENT': ['mean', 'std'],
    'CNT_DRAWINGS_CURRENT': ['mean', 'std'],
    'CNT_DRAWINGS_OTHER_CURRENT': ['mean', 'std'],
    'CNT_DRAWINGS_POS_CURRENT': ['mean', 'std'],
    'SK_DPD': ['mean', 'std'],
    'SK_DPD_DEF': ['mean', 'std'],
    'ATM_AMT_RATIO': 'mean',
    'POS_AMT_RATIO': 'mean',
    'OTH_AMT_RATIO': 'mean',
    'ATM_CNT_RATIO': 'mean',
    'POS_CNT_RATIO': 'mean',
    'OTH_CNT_RATIO': 'mean',
    'MEAN_DRAWING': 'mean',
    'CUR_PAYM_RATIO': 'mean',
    'RECIEVPRIN_RATIO': 'mean',
    'RECIEV_RATIO': 'mean',
    'BALANCE_LIMIT': 'mean',
    'DRAWING_LIMIT': 'mean',
    'PAYMENT_LIMIT': 'mean',
    'DRAWING_BALANCE': 'mean',
    'PAYMENT_BALANCE': 'mean',
    'RECIVABLE_LIMIT': 'mean',
    'RECIVABLE_BALANCE': 'mean',
    'PAYMENT_RECIVABLE': 'mean',
    'MININS_LIMIT': 'mean',
    'MININS_BALANCE': 'mean',
    'MININS_RECIVABLE': 'mean',
               }

card_agg = card.groupby('SK_ID_CURR').agg(aggregations)
card_agg.columns = pd.Index(['_'.join(('CARD', c[0], c[1].upper())) for c in card_agg.columns.tolist()])

card_agg.to_csv(locator.loc[table, 'feat_eng'])
